+++ types/weixin-app/tsconfig.json
@@ -18,6 +18,6 @@
 	},
 	"files": [
 		"index.d.ts",
-		"wx-tests.ts"
+		"weixin-app-tests.ts"
 	]
 }
