+++ types/weixin-app/tslint.json
