+++ types/weixin-app/weixin-app-tests.ts
