+++ types/weixin/tsconfig.json
@@ -18,6 +18,6 @@
 	},
 	"files": [
 		"index.d.ts",
-		"wx-app-tests.ts"
+		"weixin-tests.ts"
 	]
 }
