+++ types/weixin/tslint.json
